third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/elementRoot-expected.txt

This tests that querySelector and querySelectorAll don't search outside their root node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS root.querySelector('div') is correctNode
PASS root.querySelectorAll('div').length is 1
PASS root.querySelectorAll('div').item(0) is correctNode
PASS noChild.querySelector('div') is null
PASS noChild.querySelectorAll('div').length is 0
PASS successfullyParsed is true

TEST COMPLETE
